Move convertToMarketType to api requests helpers
